Declare window.CordSDK consistently as type ICordSDK
The type `ICordSDK` is public (i.e. part of our open-source packages), but the implementation is of the internal type CordSDK.
We do access `window.CordSDK` from the opensource code, so we need to declare it publicly, so we can only declare it as type ICordSDK.
This commit adds a static function to our private implementation (`CordSDK.get()`), which returns the global CordSDK object as type CordSDK. It throws if it hasn't been initialised (or isn't of the expected type).
This also gets rid of the "Do this better" TODOs in the code, where we used the `!` to just assume the global SDK would have been initialised.

Test Plan:
`npm run tsc-once`.
Run local server and check presence features on `https://local.cord.com:8123` work okay.

Reviewers: idevelop, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/767

monorepo-commit: 4b7dd0f13ab1245ddcb70a5243bc4cb6eab2baba