Prepare our cord-sdk packages for publishing on npm
This commit adds a couple of scripts and some boilerplate files to the `opensource/cord-sdk` folder so we can publish the packages on npm.
For each package:
* dependencies are added to the package.json, as well as some boilerplate
* a `rollup.config.js` file is added to build the package (actually these are all identical)
* a `tsconfig.json` file is added

To publish the packages:
* `npm login` to authenticate with npm
* `cd opensource/cord-sdk && npm run build` to build all packages
* and then in any or each of the subfolders in `opensource/cord-sdk/packages`: `npm publish`

When we build monorepo (both before and with this commit), our cord-sdk packages are imported as the `index.ts` file in the respective package directory. To make sure we preserve this behaviour for ourselves, this diff adds commands to delete the `dist` directories within `opensource/cord-sdk/packages` in case they exist at the beginning of our `npm run build` and `npm run watch` scripts.

Test Plan:
* run the scripts in `opensource/cord-sdk/scripts`
* publish the npm packages as described above
* run `npm run tsc-once` in monorepo to check that the main build still works
* run local server and check `https://local.cord.com:8123` and see that the collaboration features are there and work okay
* change something in one of the cord-sdk packages' source code, and see that the package gets rebuilt as well as the client code

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/806

monorepo-commit: 4ded26b900b522c77938bdc07744ede8b3de02c8