Support useCordContext in non-presence components (#958)
I noticed in the admin tool that the presence and the sidebar were
being set up in different contexts, which was a bit weird.  This was
because we use `useCordContext()` in the admin tool to set what the
context is, but the Sidebar React component didn't respect that.
Updated all the React components to pay attention to it.

Test Plan:

Loaded the local admin tool and inspected the HTML, the cord-sidebar,
the cord-presence-observer, and the cord-presence-facepile tags were
all provided the same context.

monorepo-commit: 590641566373e3cd514a24fe4ee729c9d8142758