opensouce/cord-sdk: simplify packaging boilerplate
We have `rollup.config.js` files in each package, even though they are largely the same.
Originally, I thought that it is worth having this file within each package, so that packages can be built independently and each package directory comes with all files needed to build and package.
However, this increases the amount of code we have to maintain, and it turns out there is no benefit really. Because we don't include `rollup.config.js` in the npm packages, so from the npm point of view, there is no reason to have all those files in each package folder within the repository. From a Git point of view, all the packages are distributed within one repository, so having the build infrastructure in that repo once is all we need.

This commit moves `rollup.config.js` from the various package folders to the top level of `opensource/cord-sdk`.

I tried to also move `tsconfig.json` to the top-level, but I couldn't get the generation of `.d.ts` files working correctly with that setup. To have five separate output directories which contain their respective type declarations, we need `tsconfig.json` in each of those five directories.

[skip ci]

Test Plan:
`npm run build` builds all packages. The build results are the same as before.

Reviewers: idevelop, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1025

monorepo-commit: 428f5a22a8822b240f5f8d47e0c9839e0c0e73ef