cord-sdk: add descriptions to packages
The content of the `description` field is shown as the subheadline on npm, e.g. here: https://www.npmjs.com/search?q=%40cord-sdk
We currently have no descriptions, so npm shows the first line of the README, which in our case is the HTML for the header picture. Looks terrible.

Test Plan:
I ran `scripts/update-package-json-files.mjs`, and it puts the `description` field in the right place

Reviewers: flooey, idevelop

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1058

monorepo-commit: 67c3a5237c13233f96e78270a5d0a4bd3ac93043