Move <cord-thread> to new API
This is a breaking change in the API of cord-thread, but I am
assuming that we don't have any users of the current cord-thread. Also,
the (docs)[https://docs.cord.com/components/cord-thread/] say that this
component is in BETA.

The high-level design is explained in
[Notion](https://www.notion.so/getradical/Proposal-for-new-cord-thread-API-baac42e37d124053951edd1b2f5c3e6d).

The way I decided to implement the new API boils down to:
- Given the thread externalID, fetch it from the DB. If it does not
  exist yet, then create a new EMPTY (no messages) thread in the DB.

Q: Why don't we create the ThreadEntity only once the first message is
sent? Isn't it kind of wasteful/error-prone to have empty threads in our
DB?

A: I originally started with a solution that only creates the
ThreadEntity when the first message is sent. This was achieved by having
a separate table ("platformApplicationID", "externalID", "threadID")
mapping "externalIDs" to "threadIDs" of potentially-not-yet-created
threads.
However, I found two problems with this approach.
1) it introduces a new "edge-case". Your code might have a valid
   "threadID" but you need to handle the case that the thread was not
   created yet in the "threads" table. On the other hand, our existing
   code already needs to handle the case when a thread is empty because
   all messages might have been deleted.
2) it would make features such as "(un)subscribing to/from a thread" harder to
   build. This is because our "thread_participants" table has a
   "threadID" column that "REFERENCES threads(id)". In other words, we
   will need to store information even about empty threads (such who is
   subscribed/unsubscribed) and so I decided it makes sense to create
   threadEntities even for empty threads.

Test Plan:
- went to our components test bed
  (https://local.cord.com:8179/sdk/test/)
- set the "thread-id" on cord-thread to a new id and I got a composer.
  In a new window I set the thread-id to the same id, got a composer
  and wrote a message. Sidebars and composers in both windows showed the
  new message.
- set the thread-id to an existing id and got the existing thread.

Reviewers: flooey, sven-of-cord, idevelop, HenryDavies

Reviewed By: flooey, idevelop

Pull Request: https://github.com/getcord/monorepo/pull/1059

monorepo-commit: 31a01b181992b9de6e40a260caad6d4978b0e543