Rename context to location in SDK
We're renaming the "context" concept in the SDK, after hearing feedback that it's a bit vague, and particularly confusing when used in a React app, that already has a thing called Context and it's not really the same thing as ours.

I talked it over with @flooey and we agreed that it's a change worth making, and we should make it asap rather than sleep on it, otherwise as more customers get onboarded we'll need to support deprecated props / fields for longer.

The best proposal we have so far is "location" - we think it's more concrete, semantically closer to the idea that "threads exist in locations", and it also maps nicely to the default behaviour of using the URL (aka window.location) as a fallback when a context is not explicitly defined by the developer.

Discussion: https://getcord.slack.com/archives/C0134UJLNV9/p1651747548000619

Test Plan:
- keeping all the existing context props in the testbed, components worked as before, previous threads were shown
- changing just the sidebar prop to location, and keeping the cord-thread prop as context, they correctly worked together
- changing all props to location works as expected

Reviewers: flooey, jozef-mokry

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1106

monorepo-commit: eccfe286a336f0cfe9c1979e652e7338873055a5