Expose blurring options in components (#1401)
This adds options to CordSDK.init() to enable screenshot blurring and
change the blurred screenshot display behavior.  The setup is designed
so that we could also add these to, for instance, the cord-sidebar
component in order to change them on a per-component basis, but I
don't know that it's necessary yet so I started with this.

Test Plan:

Added these to the components testbed and took an annotation.  The
screenshot was blurred for other users but not for the author, as
expected.

<img width="1403" alt="Screen Shot 2022-05-30 at 3 56 49 PM" src="https://user-images.githubusercontent.com/215991/171018235-b463a0f2-9bd5-45ae-b245-519b3ff0cb14.png">

monorepo-commit: ccc986fca8d0ba630c2f7b95100e93493fca4836