[1] Add customLocation Context property to MessageAnnotationAttachment
We want to introduce the ability for the developer to provide an arbitrary definition of the location of an annotation, which I'm  calling `customLocation` in this stack.

The existing definition of `MessageAnnotationAttachment.location` in graphQL is [unfortunately too explicitly typed](https://github.com/getcord/monorepo/blob/81c494dbc58fb4055e76d6926ca34f59d2d06909/server/src/public/common.graphql#L84-L98) to be reusable for this purpose, so I'm creating a separate field, typed as `Context`.
The expectation is that in normal usage, either `location` or `customLocation` will be defined, with the other being `null`.

Test Plan:
created a message through graphQL playground:
- when customLocation was provided, it was correctly stored in the database and returned in subsequent queries
- when customLocation was not provided it was stored as `null`

Reviewers: AlbertQM, Netceer, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1438

monorepo-commit: 200e7e9d332428aa9f291728ffa194c3c1d32502