SDK: Add showPinsOnPage prop to Sidebar to enable pins on screen
As of now, our Sidebar component renders the pins on screen (FKA hotspot/pinned annotations).
We want to make this customisable via a prop.
This diff adds `keepPinsOnPage` to the Sidebar component to do exactly this.

`showPinsOnPage` is injected into the AnnotationsOnPageContext, as we need that information in a few places.

The property is customisable only for SDK components sidebar/collaboration. The embed/extension have `showPinsOnPage` on by default.

Closes E-1875

Test Plan:
SDK testbed: added `keepPinsOnPage={false}` to the Sidebar, check I don't see the pins on the page, but I can still hover on a pill and see the arrow + pointer.

<video src="https://user-images.githubusercontent.com/33761650/172210809-4591e963-01aa-4a4a-bc14-5fee11f9421f.mov"></video>

Also checked that if I don't add `keepPinsOnPage={false}`, the pins on screen appear and work as expected.

Reviewers: idevelop

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/1452

monorepo-commit: ff133ef6176a6e1b72660fa21e118bc542ff8e2a