Annotations API: Default pin position where user clicked
Our current default position was `CENTER`, i.e. in the middle of the clicked element.
Instead, we can simply position it exactly where the user clicked. This is what we currently do for annotations.

We use this behaviour only if the user hasn't specified a renderer for that exact element.

Closes E-2088

Test Plan:
No custom renderer function, check the pin appears where I click, even for elements that are not an exact match (the row with the "random" button).

<video src="https://user-images.githubusercontent.com/33761650/174784178-61342d6c-b4d4-4c6a-add8-7b3699d516c8.mov"></video>

With a custom renderer function, that takes precedence (all pins are put at the far right of the row). However, if it's an element which doesn't have an exact match, the default is to put the pin wherever the user clicked.

<video src="https://user-images.githubusercontent.com/33761650/174784342-d3d21a02-eb92-48a2-8a50-71b3a73b353c.mov"></video>

Check what happens if annotation disappears from screen (we point to the middle of the parent element)

<video src="https://user-images.githubusercontent.com/33761650/174806969-53293fc3-4fbd-40cb-8983-a6d1f0d78351.mov"></video>

Also checked the arrow points to the edge of the screen if annotation is outside scroll.

Reviewers: idevelop, flooey

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/1579

monorepo-commit: 6450f489d2e8fe9a91bd0fbee533d27f799d8398