Annotations API: Expose redrawAnnotations method
Currently we have 3 second polling that keeps the annotations on page up-to-date.
However, this is slow. If an element disappears from the page, it can take up to 3 seconds to remove the pin.

We think allowing developers to tell Cord when to re-render annotations is a really useful feature to make the integration seamless.

Test Plan:
Implement the redraw function in the testbed.
Check the pin now scrolls smoothly with the line.

Before:

<video src="https://user-images.githubusercontent.com/33761650/175567598-c16854df-3488-4399-906c-71760f541f57.mov"></video>

After:

<video src="https://user-images.githubusercontent.com/33761650/175567685-3e1578bd-0f6c-464f-87e2-b87a7fda0d5c.mov"></video>

Reviewers: flooey, idevelop

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/1613

monorepo-commit: e51b764ea93d1aabe18832b94ecb1e2a875c8429