Thread: Add header with option & close button
When I was implementing the Thread component I realised that there's no "x" button in the Thread component, but it's very much needed.
Tom's design is to add an header with a separator on top of a floating thread. (See test plan and slack message).

I've added this logic specifically to the `<Thread />` component, and not to the ones we reuse in e.g. the sidebar.

In this first version:
- Added a `showHeader`, defaulting to `true` (happy to change), which will show the header.
- Add custom click events, so developers can run custom logic when users click on "X" button or "Resolve".

Related Docs change: https://github.com/getcord/documentation/pull/60

Test Plan:
Visit testbed, check the header is there (default is `true`). Add `showHeader={false}`, check header is gone.
Click on "X" and "Resolve", check they both fire an event (my custom logic was just to console.log the target).


<video src="https://user-images.githubusercontent.com/33761650/176219106-99c6e2db-4f82-4863-a6f0-a096e77179f5.mov"></video>

Reviewers: idevelop, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1654

monorepo-commit: 459a0fedcabe0fbec7fc1bf3b99f2f2c02a54c33