Remove All Pages from the Inbox Component
Remove `ShowAllActivity` as a property on the Inbox SDK component.  We don't want it to show up at all.  But we still want the code in sidebar to allow it, so refactor the passing down of the property value.


Test Plan:
Test on sdk\test page with Inbox component and verifying it doesn't show the All Pages page, but sidebar still does and is controllable by the ShowAllActivity prop on the Sidebar

Reviewers: yeo-yeo, flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1859

monorepo-commit: d574f5d241dde03452ca8fd1e4e13f7b93f22dd8