Fix incorrect attribute to prop conversion for boolean
When a web component attribute defined as boolean was deleted, we incorrectly parsed it to `false` when deciding the props to pass into the React component, making that component's default values not take effect. We now correctly treat `null` as meaning `undefined`.

Test Plan:
in the playground, inspect the `<cord-sidebar>` element and delete the `show-presence` attribute completely; it should still show the presence as it's default value is defined as `true`, however that wasn't the case before this fix.

Reviewers: flooey, dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/1875

monorepo-commit: ccfccab0afd9383dd9d9b834cb68248084918c92