Add inbox popup attached to inbox launcher (#1890)
Adds a popup of the inbox attached to the inbox launcher.  We decided
to have this enabled by default, since most of the time people likely
want the launcher rather than to do something else, but there's a
property to disable it.



Test Plan:

Clicked the inbox launcher on the component testbed, the inbox
correctly popped up.

monorepo-commit: c42617d92ba0aa559c071ca3c28ddf25e6aba11a