Only return SDK from CordContext after it's initialised
Currently the `CordProvider` returns the `sdk` instance ahead of it being fully initialised with the client token, which means that any downstream code that waits for the `sdk` to be available to use features like `presence` and `annotations` will hit the `Cord SDK not yet initialized` exception we throw in the getters for those SDKs.

Changing the code to introduce a new boolean state variable so that we only return the `sdk` instance once it's finished initialising.

Test Plan:
before this change, a hook that tried to call `sdk.presence.setPresent` as soon as the `sdk` instance became available on `CordContext` would hit the exception; now, it doesn't;


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/1894

monorepo-commit: bbf6ab15bd7e6da044fa7cae9322564feffbceda