Set up ability to forward refs to SDK React components
In some cases we'll want to be able to call methods explicitly defined on a custom element such as `<cord-sidebar>`.

In this PR I'm defining a basic `HTMLCordSidebar` interface that defines a single method named `startComposer()`.

In order to call this from the developer code, I'm creating the possibility to pass a `ref` object to `<Sidebar>`. This ref will be updated with the instance of the element, but only after the custom element class itself is registered and initialises the element (otherwise the ref correctly points to the DOM element but the custom element instance backing it isn't there, so the various custom methods and attributes would not exist).


Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/1977

monorepo-commit: 5ebc227592b982922b435778f9097a81214afcd5