AnnotationPins component
See https://linear.app/cord/issue/E-2329/annotations-component for more details.
Lots of users have expressed the need for a component that allows to create annotations and floating threads together.
As of now, the way we've told client to implement this is: 1. create your own annotation (e.g. draw the pin on a canvas); 2. render Cord Thread next to it.

The component added in this PR aims to give users the power of our annotations. They will be able to use our annotation flow (which comes with the little pointer following your cursor), the annotations pointers will stick to the right place, etc.


Reviewers: idevelop, flooey

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/1968

monorepo-commit: 82ee640fc7d1d752e2aaa9a57dacf99b4e7ab942