AnnotationPins: Add `startAnnotating` function to programmatically start the annotation flow
Users expressed the need for a way to start the annotation flow programmatically for a long time.
Introducing it in this PR, following the pattern we already have.

Test Plan:
Add a button which calls `startAnnotating`, check it correctly starts the annotation flow.


Reviewers: idevelop, flooey

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/2175

monorepo-commit: dc076251cd5eee0dac0f1d6d22f49ea24fe1c4c6