Rename AnchoredThreads -> FloatingThreads (#2277)
Rename AnchoredThreads component to FloatingThreads.  The name
AnchoredThreads is left as a backwards compatibility name in three
places: the two exports from @cord-sdk/react and one type export used
by the canvas sample app.



Test Plan:

Loaded the floating threads component on the testbed, clicked a
pointer to show a thread, closed the thread, clicked the button to
leave an annotation, and sent a message in the resulting thread; all
worked.

monorepo-commit: b467d7ef2c6ed71968257f3d861f1903762c64ef