Add relative coordinates of the annotation target to `useCordAnnotationRenderer`
This adds the x,y coordinates of the element as relative positions when we call the developer back in `useCordAnnotationRenderer` to get the proper location for display of an annotation pin.  This allows the developer discretion to use the original click if they so choose.  The coordinates are given as relative to the original size of the element.  This allows for them to move should the element resize.  They will be an x,y pair between [0,1].


Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/2320

monorepo-commit: 31e9157aced9c96d8a4bbf4714f0b2638c61089b