Add support for annotating only targeted areas
This introduces the concept of Target Annotations Mode to the SDK. In this mode, annotations may only be placed on elements decorated with `CORD_ANNOTATION_LOCATION_DATA_ATTRIBUTE`.  Only those elements can be selected for an annotation. Other places on the page may not be selected.

The default mode for Annotations will still be to allow them anywhere on the page.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/2355

monorepo-commit: c1ac0689a2db3b11148ba90cc28eea7da2684c20