Sidebar: pass width into onOpen callback
This can be hard for client code to measure, so let's pass it
in.

Instead of being added as a parameter directly, add an object as the
first parameter to `onOpen` and add `width` as an optional field in that
object. This is to help with backwards-compat in the future: if we ever
want to remove `width`, we can do so (since the field is optional), but
if it were a parameter directly, it would be impossible to *ever* remove
since it would shift the other parameters "up" one slot.


Test Plan:
In `sdk/test/index.tsx`, update the `Sidebar`'s `onOpen`
callback to print out its first argument. See the correct width passed
in. Change `--cord-sidebar-width`, see that reflected in the log --
clamped at the `min-width` and `max-width` we set.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/2564

monorepo-commit: 82d805196e1455edad4845cb481ead9cbbc4c3a7