Add threadName attribute
Customers have requested being able to set the thread name directly on a
component, instead of us setting it automatically based on the page
title (or falling back to the URL). Add that attribute and thread it
through to where the thread is created. If the thread already exists,
the attribute is ignored.

The following components get the new attribute (since they can create
threads):

- `cord-floating-threads`
- `cord-sidebar`
- `cord-thread`

The semantics for `cord-thread` are potentially slightly awkward -- the
thread is created when the component is first rendered, not when the
first message is sent to it. We can always improve later if needed.


Test Plan:
For `cord-thread`, it made sense to add a field into `sdk/test`;
verified that threads without names still fall back to URL; threads with
names get that name:

https://user-images.githubusercontent.com/575697/194070974-0bd1934d-6be6-4243-a9d9-979aae1f75d4.mov

For the two other components, it doesn't really make sense to add such a
field, so I manually tested that by defaul they still fall back to the
URL, but a specified `threadName` properly set the name for any new
threads.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/2606

monorepo-commit: 4b1942202d63eabb588c0f32f72c455e2d4316bd