api-types: add property orders to JSON schema
This commit switches on the `propOrder` setting, which adds `propertyOrder` fields to the JSON schemas.
It also changes the order of some properties in `typeDefinitions.ts` to match the order we currently have in the API documentation.
I am looking at generating the API documentation from this schema, and I want to have a defined order for the props.
While JavaScript (or at least Node.js) preserves the order in which items appear in an object, other languages may not. Our documentation is generated by Jekyll, which is implemented in Ruby, and I have no idea if there is a guarantee that when iterating over fields, the order is the same as it was in the JSON file. Either way, explicit is better than implicit, so let's just have this extra field in the JSON schemas, which defines the canonical order of fields.

Test Plan:
code review

Reviewers: flooey, mhtien

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/2620

monorepo-commit: 4604e89716efe8df74d5e76f385eea7d7a901f55