Introduce data-cord-annotation-allowed attribute
A developer can now decorate their DOM with `data-cord-annotation-allowed` attribute, which accepts boolean `true` or `false` values (case insensitive).

If an element is decorated with `data-cord-annotation-allowed=false`, then that element cannot be annotated: when the user is in annotation mode and hovers over the element, the pin disappears and click events do nothing; every other element is by default allowed to be annotated; this is useful when the majority of the page should be allowed to be annotated, except for a specific part of the DOM, like a floating menu.

If an element is decorated with `data-cord-annotation-allowed=true`, this will override any `false` definitions higher up the parent chain. This is useful when the vast majority of the page should not be allowed to be annotated, except for a specific part of the DOM, like a text editor inside a CMS


Reviewers: jwatzman, dmmiller, flooey

Reviewed By: jwatzman, dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/2650

monorepo-commit: be2e8e157b9c1fdb21d872bb1bab52e3d8e14b6b