SDK(beta): SelectionComments component
Adding the TextComments component to the SDK beta package.

This PR contains all the boilerplate code to publish our component within our js package.
There is one particularity about the TextComments component. Because we wrap the
areas that are allowed to be annotated with the actual web component, when we render
it - and thus put it within a shadow root - it takes precedence against its children. For that
reason, @idevelop suggested that we wrap the component in another div and put that into
the shadow DOM, so its children aren't affected. Apart from that, the other issue was that now
we did not have access to the children in order to be able to attach listeners. So when we
render the component, we also add attributes to its children, so we can find them and attach
the listeners to them later.

An alternative approach for the second issue was to pass the children down as props to the
component, which was quite cumbersome to do and would also render the children within the
shadow DOM, which meant that they would lose their existing styles.

Test Plan:
In the testbed, I selected a piece of text. Then scrolled all the way down to find the
"Comment" button. Because we attached the button as a child to the body, and there is no
logic to render its position yet, it renders at the bottom of the page.


https://user-images.githubusercontent.com/102963329/199999235-161eb3ea-b7ad-4ca4-becb-25d451f0d0e5.mov


Reviewers: idevelop, AlbertQM, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/2750

monorepo-commit: 78744f309d02499df1fc803520d00c6953bcbb81