Rendering composer within the correct place when a web component wraps the DOM
I have edited the code which grabs the web component element when
we initialise it. As mentioned in https://github.com/getcord/monorepo/pull/2750, because the SelectionComments component wraps other DOM elements, we had to do a workaround in order for the children of the component to be positioned in the light DOM and not actually disappear. This different wrapping though causes the slot + portal trick (https://github.com/getcord/monorepo/pull/1190) we have in place for the composer to break. The root of the portal is pointing to the parent component, and not the wrapper, which makes the composer render outside the SelectionComments thread. From now on, all components which wrap the client code will have the data-cord-component-wraps-dom attribute, so we can correctly point to the wrapper and render the composer in the correct place.

@AlbertQM Made this very helpful diagram:
image

Test Plan:
Rendered a composer in the SelectionComments component (using #2773) and the cursor was typing
in the correct location.

https://user-images.githubusercontent.com/102963329/200901902-6fce4119-cbd4-4c77-a683-c7dc0ee0068e.mov


Reviewers: flooey, AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/2791

monorepo-commit: a5cb500fbac21827b807c7a3d89e675e5b8b1187