Thread component: Add `autofocus`
Some clients have raised the need for our Thread component to be in focus whenever they render it on the page.
We already have a prop that does exactly that, I'm now exposing it in the `Thread` component.

Test Plan:
Add `autofocus` to our testbed's Thread. Check that when I open the page, the `Thread` has focus.


Reviewers: flooey, jwatzman, idevelop

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/3211

monorepo-commit: 161e271b8f2faba220bb5a1f9330b7f9a7455101