Add `style` attribute to React API components
This allows the `style` attribute to be specified in our React API;
there's even a comment here that we maybe wanted to do this but never
got around to it.

Since `style` takes a React-specific type, I couldn't add it into the
type in the types package. However, since that seemed to be
React-specific anyway, I moved that type into the React package -- as
well as deleting a couple of vestigial-looking types next to it.

Doing that is technically a backwards-compat break. Given these types
look like utility types for our React API anyway, I'd be pretty
surprised if a customer were relying on them in practise.


Test Plan:
Add `style={{ border: '4px solid purple' }}` to a `Thread` in the SDK
testbed. See a purple border.

Reviewers: idevelop, flooey

Reviewed By: idevelop, flooey

Pull Request: https://github.com/getcord/monorepo/pull/3282

monorepo-commit: fb7fa99232be9f3e5df3d81dc6c62573c475991b