Add localStorage key to override SDK script URL
This uses a similar idea to the loading times logging, to add a
localStorage key which can be manually set in order to override the SDK
script location. While there is already a `cordScriptUrl` prop which
controls this, using the localStorage key allows easily playing with
this without editing/updating the main app.


Reviewers: flooey, idevelop

Reviewed By: idevelop

Pull Request: https://github.com/getcord/monorepo/pull/3318

monorepo-commit: 817115b335a178226424dbc74a524082002b9844