Create skeleton of Composer component (#3396)
This supplies a first version of a separate component that is just a
composer.  Depending on the value of thread-id, it either creates new
threads or adds messages to an existing thread.

The component doesn't react properly to width/height styling yet, nor
does it issue any events, which we might want.


Test Plan:

Loaded testbed and sent messages, added attachments, and @-mentioned
users, it all worked properly.

monorepo-commit: a4aa3a1369c496909a7b438457b2a8d29a76ec71