Upgrading the typescript-json-schema package
I attempted to make a changes in our REST api validation logic and
realised that the script to generate the new schema was not working. After
investigating with @mhtien and @sven-of-cord we found that upgrading the
library to the latest version, it correctly generates the API schema.

Test Plan:
`cd opensource/sdk-js/packages/api-types` -> `node generate.mjs`


Reviewers: flooey, sven-of-cord, mhtien

Reviewed By: sven-of-cord

Pull Request: https://github.com/getcord/monorepo/pull/3586

monorepo-commit: a3abe03cf58d260ec66493277cbb873cadea0c67