Allowing for the composer to always be expanded (via a prop) in the thread component
From the cordathon muddles, we got:

"I'd also like to toggle the 'expanded' composer on by default, rather than have it expand when active"

This PR exposes a prop for the Thread component to allow for having the composer expanded always.

I have also added a toggle for this prop in the testbed.

Test Plan:
https://user-images.githubusercontent.com/102963329/218327119-0cd8e7b6-06e5-48e6-832b-d894a2784a3a.mov


Reviewers: mhtien, flooey

Reviewed By: mhtien

Pull Request: https://github.com/getcord/monorepo/pull/3607

monorepo-commit: d55634d89a2d2d0adb5a8c9f028a36267a89c83c