Allowing a null value for profile_picture_url in our REST API
We want to allow users to revert their profile pictures back to null (default letter of the name will show).

This was requested by customers and was also identified as a bug in the "Cordathon"

This PR modifies the REST API types to allow for a nullable profile_picture_url value. It also modifies the
user mutator to update with a null value.

Test Plan:
Created a dummy user (via the REST API) with a sample photo. Initialised a Cord app with this user and saw the
correct profile picture URL. Then made another API call with the `profile_picture_url`: null, json value. The picture in the
Cord app reverted to the first letter of the user's name.


Reviewers: mhtien, jwatzman, sven-of-cord, flooey

Reviewed By: jwatzman, sven-of-cord

Pull Request: https://github.com/getcord/monorepo/pull/3587

monorepo-commit: 9911c0a6c95d8607207c1a85eeeb71977e65dabd