Change JS presence API to exact match by default (#3714)
We've found that most developers expect our APIs to use exact matching
on locations by default, so change the presence API to use exact
matching by default, but with a `partial_match` option to enable
partial matching.  Also update useCordPresentUsers to do the same
thing.

This is safe because we've verified that nobody is calling these APIs
except through the presence web components, and those are updated in
this diff, so users will get either one or the other.

We unfortunately can't change the default for the web components,
since they're in use by lots of people, but I've added the
partialMatch argument in addition to the existing exactMatch one, so
that people can use the same argument names in both places.



Test Plan:

npm run build

Loaded the testbed, the table demo (which uses useCordPresentUsers)
still works properly.

monorepo-commit: 6aec457e366b047a90d3d1c59a5165596a369a1b