ThreadList: Add onThreadHover
Discussed in HOHO with Nimrod, Jg and Adam.
A client has expressed the need to highlight pins on screens from the ThreadList.
In the sidebar this is done with our arrows. The ThreadList has nothing.

We decided to expose an on hover event, so that developers can at least do something about it (e.g. embiggen the pin when hovering on the thread).

When implementing this API, I noticed that it's impossible to e.g. implement a custom hover styling on the threads inside ThreadList. We should fix this.

Test Plan:
Added the callback to our testbed, check IDs get logged correctly on hover, and that moving the cursor on the thread's menu doesn't trigger extra events. With the thread's action menu open, click on the thread and see no "mouseleave" event. Reopen the menu, click outside the thread, see "mouseleave" event.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/3827

monorepo-commit: 2f40156c28572ddbd5cbf23ea656f44cd7aeb715