Add events to FloatingThread when an annotation is started, finished, and cancelled
This adds three events to FloatingThreads.  It will now fire onStart, onFinish, and onCancel during usage.  onStart will be called when a user clicks the button to begin an annotation.  onFinish will be called when an annotation is completed.  This is after the user has entered their text in the composer and sent it.  onCancel will be called if the user hits escape (or cancel) during the pin placement in addition to canceling the composer.


Test Plan: Verified in local testbed with console logging

Reviewers: jozef-mokry, mhtien, flooey, yeo-yeo

Reviewed By: flooey, yeo-yeo

Pull Request: https://github.com/getcord/monorepo/pull/3905

monorepo-commit: cd4aa92edba08665b1b34d20f3e8bca4dd72ffa0