Add clientAuthToken, userID, and organizationID to CordContext (#3855)
Summary: In building with Cord, I repeatedly needed to understand who
the user is currently acting as. It's super freaking easy to create
threads in the wrong org when you don't know. Until this PR, there's
actually no way to know what token the SDK is using nor which userID or
orgID it's seeing. I believe even this change doesn't fully scratch the
itch because the SDK still takes a couple ticks to actually switch over,
but this is really freaking useful for devs trying to work with our
stuff. I hit this problem with the sample apps when doing the docs site
rewrite and when working on the puzzle site. I'm sure any customer of
ours doing org switching will also hit these issues.

A lot of what's here is a port of work that My Hoa did to extract the
org ID from the token, but that work was buried in a hook in the demo
apps. This would be useful across all our callers.

monorepo-commit: d62c845829c2d811aff409b29ef26f8bd3cfacd4