Add cancelComment method to FloatingThreads
This adds the ability to cancel an annotation programatically for FloatingThreads.  Doing so will cancel the current annotation if you are still selecting the pin placement or have the composer up.


Test Plan:
Added a new button in testbed that allows you to enter comment mode and stay in it until you exit it by cancelling or clicking the button to exit

Reviewers: AlbertQM, yeo-yeo, mhtien, flooey

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/3915

monorepo-commit: 433f29080a4763698fa452daf49024af50d2ff0f